refactor: move Bugzilla class to lib_bugzilla.py
Move the Bugzilla class, BugzillaAPIError exception, and
_bugzilla_error_body helper from mcp_utils.py to a new
lib_bugzilla.py module for better separation of concerns.

- mcp_utils.py now only contains logging setup (ColorFormatter,
  mcp_log) and utility functions (is_textual, safe_filename)
- mcp_utils.py uses __getattr__ to lazily re-export Bugzilla
  and BugzillaAPIError for backward compatibility, avoiding
  circular imports
- server.py imports Bugzilla directly from lib_bugzilla
- test_mcp_utils.py imports Bugzilla directly from lib_bugzilla
